Travel Settings Screens: Edit a Travel and manage Participants (#40)
* feat(travelsettings): create skeleton screen

Create skeleton UI for EditTravelSettingsScreen.kt with a ViewModel. Not functional yet.
Relates to #10

* fix(travel-repository): add toMap method and correct time fields (#39)

feat(travel-repository): add toMap method and correct time fields

- Added `toMap` method to `TravelRepositoryFirestore` to facilitate serialization.
- Corrected `endTime` and `startTime` fields in `documentToTravel` method.
- Updated unit tests to reflect these changes and ensure functionality.

* feat(TravelSettings): Add draft Settings Screen

Add a basic, main screen containing the required
travel information and buttons. No buttons except saving settings are working.
This is required for #10 and is needed for the
participant roles screen that is to be implemented.

* chore: typealias for fsUid

Add typealias to clarify the use of fsUid in functions and objects.

* feat: Add participantList and userInfo

Add extra screen to list all participants of a travel.
Add UserInfo data class and Add retrieval logic to download it from
Firestore.
TODO: Finish up retrieval and upload

* feat(TravelSettings): Add fetch logic and display

Add firestore logic to retrieve user information to be displayed.
Retrieval is correct, but needs to be done in ViewModel as a flow.

* chore(TravelSettings): remove fields from UserInfo

Remove fields from UserInfo that are too complex
for now and unused.

* feat: Add ViewModel user info retrieval

Add userInfo fetching that actually works when one selects
participantList. Modify ViewModel and ParticipantListScreen to
effectively avoid redundant calls.

* feat: Add user edit and deletion popup

Implement display logic to change roles of participants
as well as removing them from the Travel.
No actual data is modified.

* chore(travelsetting): Rename email button

Rename email button in order to correspond to the actual use.
This button is simply supposed to paste a link into the clipboard of the
user.

* chore(travelsetting): ktfmt format for CI

Ktfmt format EditTravelSettingsScreen.kt in order for the CI to succeed.

* feat(TravelSettings): Add draft Settings Screen

Add a basic, main screen containing the required
travel information and buttons. No buttons except saving settings are working.
This is required for #10 and is needed for the
participant roles screen that is to be implemented.

* chore: typealias for fsUid

Add typealias to clarify the use of fsUid in functions and objects.

* feat: Add participantList and userInfo

Add extra screen to list all participants of a travel.
Add UserInfo data class and Add retrieval logic to download it from
Firestore.
TODO: Finish up retrieval and upload

* feat(TravelSettings): Add fetch logic and display

Add firestore logic to retrieve user information to be displayed.
Retrieval is correct, but needs to be done in ViewModel as a flow.

* chore(TravelSettings): remove fields from UserInfo

Remove fields from UserInfo that are too complex
for now and unused.

* feat: Add ViewModel user info retrieval

Add userInfo fetching that actually works when one selects
participantList. Modify ViewModel and ParticipantListScreen to
effectively avoid redundant calls.

* feat: Add user edit and deletion popup

Implement display logic to change roles of participants
as well as removing them from the Travel.
No actual data is modified.

* chore(travelsetting): Rename email button

Rename email button in order to correspond to the actual use.
This button is simply supposed to paste a link into the clipboard of the
user.

* chore(travelsetting): ktfmt format for CI

Ktfmt format EditTravelSettingsScreen.kt in order for the CI to succeed.

* chore(travelsetting): delete unnecessary comments

Remove comments left in the code that serve no practical purpose

* docs(travelsetting): add documentation for method

Add documentation to documentToUserInfo since it was requested in PR

* refactor(travelsettings): extract into composable

Extract chunks of logic into smaller composable to enhance readability.

* docs(travelsettings): document composables

Give a bit more documentation for the newly extracted composables

* chore(travelsettings): ktfmt and minor fixes

Ktfmt compliant files and fixed fetchAllParticipant to compare if the
list of participant AND their roles have changed before querying.

* feat(travelsettings): Add correct save parsing

Add correct fields to be edited as wall as correct parsing.
Save edited information to firestore correctly.

* feat(travelsettings): Add user removal logic

Implements the removal and update to firestore of user removals.
Buggy as of now.

* chore(travelsettings): ktfmt format file

ParticipantListScreen.kt somehow was not formatted according to ktfmt
format. This is needed for the CI to work

* feat(travelsettings): Add user role change logic

Implement user role change and synchronise role change with Firestore.

* test(travelsetting): Unit & UI test setting screen

Add Unit tests for firestore functions, Timestamp parser, UserInfo.
Add UI tests for EditTravelSettingsScreen and for ParticipantListScreen.
Add testtags to composables.
These acheive atleast 80% coverage and but can be refined.

* fix(travelsetting): hard set a locale for CI

Hard fix a locale in the parseDateToTimestamp function. This may prevent
CI from failing the test.

* chore(travelsetting): cleanup and small test

Clean up MainActivity for merging into main. Add useless pointless test
to increase coverage. Documented disabled test.

* chore(travelsetting): typo in variable name

Fix typo in the newTravel creation after editing.

* Revert "chore(travelsetting): typo in variable name"

This reverts commit 0761d1e38de35939e5b84807129d3abcc2fe503f.

* fix(travelsetting): typo in variable name

Fix typo in the newTravel creation after editing.

* fix(travelsetting): make textfield scrollable

Make Textfields scrollable and change tests so that the CI will not fail
because of display issues.

* chore(travelsetting): ktfmt format file

Reformat EditTravelSettingsScreen.kt to pass the ktfmt format check.

* fix(travelsetting): seeing why the test fail in CI

Add checks for components to check why the CI assert fail while working
locally.

* fix(travelsetting): identify and fix CI

Changing the tests to identify why the CI doesn't pass.

* chore(travelsetting): ktfmt format

It has to have the right ktfmt format...

* fix(travelsetting): triggering CI

Since the CI gives 0 indications of the failure locations, we have to
test stuff

* fix(travelsetting): CI scroll to bottom buttons

Adding .performScrollTo() to save and delete button because CI has
a very small screen :(

---------

Co-authored-by: appellet <160597306+appellet@users.noreply.github.com>