Feature : Travel activities (#47)
* feat: Adding the Activity class, the repositories, model view. Adding also the screens to add an activity to a travel, and the listing of all activities

* tests: Adding tests for the activity logic, model view and UI.
fix: correcting some signatures and some Strings due to copy paste (replacing Events to Activity)

* docs: Adding documentation to activity model view, repository and screens.
fix: making the model view log the exception in case on failure. It also shows a toast in case on failure or success.refactor: putting the parsing of the date in AddActivity screen in a helper function and replacing getActivities by getAllActivities

* docs: Adding documentation to activity model view, repository and screens.
fix: making the model view log the exception in case on failure. It also shows a toast in case on failure or success.
refactor: putting the parsing of the date in AddActivity screen in a helper function and replacing getActivities by getAllActivities

* fix: Correcting some typos (ActivityModelView to ActivityViewModel), setting default value of selected activity to null, and adding the name of the functions (onSuccess or onFailure) if the functions where they are called, for better readability.
docs: adding documentation for the class ActivityViewModel

* style: Applying the ktfmt format

* refactor: putting the dateVisualTransformer for the date text field to a separate file and removing useless comments

* refactor: putting the dateVisualTransformer for the date text field to a separate file and removing useless comments
docs: adding documentation to the dateVisualTransformer file and the addActivityScreen

* feat: adding top bar app on the add travel screen

* style: applying ktfmtFormat

---------

Co-authored-by: Victor Rydmark <160653991+RemIsMyWaifuu@users.noreply.github.com>