Ci: Release build for CI  (#94)
* Ci: Release build for CI and minimize size

Modified apk workflow to build a release and print SHA-1 fingerprint.
Fingerprint is needed for Google Sign-in, and proguard-rules.pro and
build.gradle.kts minimize the size of the resulting apk.

* Ci: improve passing of signing key to CI

Pass keystore, keystorepassword, keyalias, keypassword as environnment
variables. This is cleaner, and allows the signing profile to be only
active whenever an APK build is launched.
Fix failing CI by not hard coding key paths.