Latest activities reminder banner #80 (#105)
* feat(reminder): Basic reminder of activities

Add basic overlay that lists activities that are due in less than 24h.

* feat(reminder): Add outline and resize banner

Add outline to the reminder banner when tapping on it.
This makes the outline of the dismiss button more visible.

* chore(reminder): Remove unused import

Remove an unused import that is unnecessary

* test(reminder): Add tests for banner composable

Add UI tests to acheive sufficient coverage for sonar cloud.

* chore: Add extra test and style improvements

Add extra test case for activities not due soon.
Change UI behaviour on banner if no activities are due by calling
onDismiss to show more clearly the logic flow.

* chore: Add extra test and style improvements

Add extra test case for activities not due soon.
Change UI behaviour on banner if no activities are due by calling
onDismiss to show more clearly the logic flow.

* fix(reminder): Fix timedelta to seconds instead

Change the timedelta duration to the correct amount.
Previous value was in milliseconds not seconds.

* test(reminder): Add time delta tests

Add a test case for events closer to 24 hours. This ensure correct
behaviour for these edge cases that might not be spotted otherwise.