Feature/notification UI with refactor (#145)
* feat(navigation): create a path to notifications on BottomNavigationMenu

Create a button on the bottom navigation menu to access to the notification view

* feat(notification): fetch notifications from Firebase

Added a function documentToNotification that deserialize the notifications on Firebase in a Notification object. Printed the notification of a specific user on the screen in NotificationScreen

* feat: implement the sending of invitation to a travel

Added the implementation to send a request to someone to join a travel and also the answer.

* feat: add user to travel when accept invitation

Implemented add a user to a travel when clic on "ACCEPT" button after another user invit him to join the travel

* feat: add notification for update role

Implemented the notification when the role of a user is updated

* feat: update notifications when delete all

Implemented the button to delete all notifications directly when push the button.

* chore: formatting files with ktfmt

Formatted files to improve the consistency of the project

* refactor: apply big refactor to notification feature

Addapted the big refactor of Firebase on the notification feature.

* chore: downgrade agp version for compatibility

Downgraded the version of the agp library for compatibility with all member of the groups. The new version of Android Studio suggest to upgrade, but we keep the old version.

* fix: fetch a travel with its fsuid after refactore

Modified the getter function for travel form the Firestore database after big refactoring.

* chore: delete simulation to resolve main conflict

Deleted the simulation test notification file to resolve conflict with the main branch.

* fix: call getTravelById for notification

Implemented getTravelById to retrieve a specific travel from its fsuid to use it for sending notifications.

* feat: invitation cannot be done to ourself

Controled that an invitation cannot be send by ourself in the Edition of travel. Esnure compatibility with ProfileViewModel

* feat: update travels to user2 when accept invitation

Updated the list of travels when accepted an invitation from another user

* feat: implement navigation to go to travel notification

Implemented the path from the notification of a specific travel to the activities view of the travel.

* test: update unit test Notification model for coverage

Updated NotificationViewModelUnitTest, NotificationContainerValidDataUnitTest.kt and NotificationRepositoryUnitTest for coverage and adaptation purpose to new modifications of Notification UI.

* test: update ParticipantListScreenTest with new Notification model

Added notification and profile view model in participant list screen to adapt tests with the new stuff

* refactor: rename NotificationRepository as NotificationRepositoryFirestore

Refactored the name of NotificationRepository after create an interface NotificationRepository that is implemented by NotificationRepositoryFirestore

* test(notification-ui): add NotificationsScreen categories tests

Added Unit Test on notifications Screen time categories that dispatch notifications on different screen

* refactor(notification): simplify NotificationItem

Added helper functions to simplify NotificationItem. This simplify also the test of the function

* chore: format files with ktfmt to optimize import

* test: delete a wrong test for NotificationItem

Deleted handleInvitationResponse because wasn't implemented correctly and doesn't pass the CI

* test: add test getTravelById in travel view model

Added a test to a new function getTravelViewModel in ListTravelViewModel

* test: Add test for NotificationScreen coverage

Add extra button presses and mock notifications to increase coverage.

* fix: Revert deleted library

Try to fix CI by restoring library that was seemingly useless.

* fix: Fix broken test for EditTravelScreenTest.kt

Fix incorrect test because leading to CI failure.

* test: Modify EditTravelScreenTest.kt AddUser flow

Modify EditTravelScreenTest.kt to test behaviour of new AddUser logic.

* chore: Remove commented out lines

Remove commented out lines from unused test.

* test: Adding test for profile repository function getFsUidByEmail and travelRepository function getTravelById

* fix: correcting the succeeding function of getFsUidByEmailTestSucceedsWithEmptyList with 'it' and not 'profile.fsUid'

* test(notification): add tests for deleteAllNotification

Added a test in deleteAllNotificationForUser of NotificationRepository

* test: add getFsUidByEmail for ProfileRepository

This test was deleted and rewrite in this commit

* fix: update main changes in the branch

Implemented new function in mock repository.

* fix: adapt main to branch

Added getFsuidByEmail in ProfileRepository since it was deleted during the merge.

* fix: adapt tests on deleteAllNotification

Added a Failure listener on delete function of repository. Needed to adapt tests in consequence.

* fix: adapt NotificationRepository to end2end tests

Used FirebasePaths in NotificationRepositoryFirestore to be adapted to end2end tests

* fix: change setContent in all tests

Moved setContent in all tests instead of setUp function

* fix: a few inaccuracies in NotificationRepository

Added documentation in NotificationRepository and correct different inaccuracies in NotificationRepositoryFirestore.

* chore: format code with ktmft

Fromated all project with ktmft

* fix: delete reload notification on screen

Deleted the update of notifications on the screen once click on the button to delete all notification.

* fix: change name with username in notificationItem

Updated the notification message with the username instead of the name by default.

* fix: update requested changes in NotificationRepository

Updated requested changes in NotificationRepositoryFirestory.

* test: Completing the test suite for adding a participant with notifications

* test: Adding a do answer to get a notification uid when clicking on the add suer button, even if the function is not called

This helps us indeed verify that the function is never called. which makes the test valid

* test: Verifying that the function getTravelById of travel repository is called when clicking on the button ACCEPT or DECLINE

---------

Co-authored-by: RemIsMyWaifuu <victor.rydmark@epfl.ch>
Co-authored-by: RemIsMyWaifuu <160653991+RemIsMyWaifuu@users.noreply.github.com>
Co-authored-by: yassine04e <yassine.elgraoui@epfl.ch>
Co-authored-by: Yassine El graoui <164810676+yassine04e@users.noreply.github.com>