Improve UI and logical review on EditTravelSettings (#223)
* feat: add localisation menu instead of coordinates

Deleted the 2 coordinates field for localisation and replace with query search menu for localisation. Also implemented a button to access to participant list instead of add directly someone.

* feat: add fetching participants info

Added the fetch of participants infos when click on the Manage Participants button. This ensure to have updated data on the screen

* fix: changed go back button participants list

Changed the button back button in ParticipantListScreen to avoid infinit loop back between participantsListScreen and EditTravelSettings

* fix: change own role

Fixed the crash when someone try to change the own role that was causing by notification. Now, the notification is send only if the receiver is not the sender.

* feat: put down button after click

Implemented the automatic retract buttons after clicking on one.

* fix: show dialog on button click

Fixed the button to add a user on a travel that show now the dialog.

* test: adapt test ParticipantsListScreen

Adapted the tests of participant list screen to match with the new UI

* test: adapt add User tests

Adapted the tests to add a participant to a travel since the button move from EditTravelSettings to ParticipantsListScreen

* test: apadt tests add user

Moved tests add user from EditTravelSettings to ParticipantsListScreen.

* chore: delete add user in editTravel

Deleted the dialog add user in edit travel to avoid duplication and unused code.

* test: dropdown menu for query location

Added tests for query a location in the EditTravelSettings with the dropdown menu

* test: adapt with merge main

Adapted tests with newest version of main

* feat: define style in Participant Column

Defined the style in participant column to be consistent with the remaining UI of the app

* feat: uniform field text width

Changed the width of the field text to be consistent with the other screen.