Modify UI and consistensy across the app (#252)
* feature: change UI travel card

Modify font and exchange Title and Date in the card in the screen TravelList

* feature: add padding field in edit travel

Added padding from horizontal border to field of EditTravelSettings screen

* fix: ajdust test for no activity text

Adjusted the tests of PageSwipeTest for no activity message in the middle of the screen

* feature: add time field in Edit activity

Added the possibility to change the time in an activity to be coherant with the creation and definition of an activitiy

* feature: change location activity

Added the possibility to change the location of an activity with query location.

* test: adapted tests for location query

Adapted test for location query in edit activity screen

* fix: test e2e DocumentUpload

Fixed the tag emptyTravel with the new tags

* feature: generalize query location field

Added a general function composable to query a location for an Activity. It's used in AddActivity and EditActivity.

* feature: generalize query location field

Added a general function composable to query a location for an Activity. It's used in AddActivity and EditActivity.

* feature: generalize date and time field

Generalized time and date field with a call to DateTimeInputField.

* chore: add upper case to LocationQueryField.kt

Changed as requested the name of the composable with an upper case LocationQueryField.kt.